Version: 1.0
Standardised a bit of the different font sizes used in this program

Created a AbstractViewComponent class which makes all of its subclasses
have a border and a title when it is being displayed. It was required to
add some abstraction to the code, because the actually 'gameplay' I am
still a bit uncertain on how to implement it. I try to make the GUI
components more modular, in order to make it easier to interface with
the controller modules. And its also fun to refactor your code !! :D