Version 1.0
Started with placing the tokens on the board and implementing the
'Referee' (while it seems relevant)

I want to make the players in the game be modelled as a link listed,
because that would enforce which turn it was. It may have
overcomplicated stuff, but I believe that if one can reflect programming
models to real-life it makes it easier in the end. Let us hope so.

Must do:
- Get some gameplay going
- Finish up with GUI (i.e. Displaying Chance card and Property cards)
- Revise the Monopoly game rules
- TEST UNITS!!!!!

====

I want to do:
- Use image processing to detect the property lots on the board.
- Make better icons
- Make splash screen (neat up the GUI)